feat: it can store a plan with a cap

+++ database/migrations/2023_07_01_083950_create_plans_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('name');
             $table->date('start_date');
             $table->date('end_date')->nullable();
-            $table->integer('target_amount_per_month');
+            $table->bigInteger('target_amount_per_month');
             $table->string('target_variable');
             $table->string('payout_frequency');
             $table->foreignIdFor(Admin::class, 'creator_id');
